+++ Savings-API/Services/SavingsService.cs
@@ -9,13 +9,11 @@ namespace Savings_API.Services
 {
     public interface ISavingsService 
     {
-        public IList<SavingVm> GetAllSavings();
-        public IList<SavingVm> GetSavingsForYear(int year);
-        public IList<SavingVm> GetSavingsForMonth(int year, int month);
+        public IList<SavingVm> GetSavings(SavingsFilterDto filter);
         public SavingVm? GetSavingVm(int savingId);
         public Saving? GetSaving(int savingId);
-        public Task<Saving> AddSaving(AddOrEditSavingDto dto);
-        public Task<Saving> UpdateSaving(int savingId, AddOrEditSavingDto dto);
+        public Task<SavingVm> AddSaving(AddOrEditSavingDto dto);
+        public Task<SavingVm> UpdateSaving(int savingId, AddOrEditSavingDto dto);
         public Task DeleteSaving(int savingId);
     }
 
@@ -30,36 +28,36 @@ namespace Savings_API.Services
             _mapper = mapper;
         }
 
-        public IList<SavingVm> GetAllSavings()
+        public IList<SavingVm> GetSavings(SavingsFilterDto filter)
         {
-            List<Saving> savings = _dbContext.Savings.Include(s => s.User).Include(s => s.Goal).AsNoTracking().ToList();
+            var query = _dbContext.Savings.AsQueryable();
+
+            if (filter.DateFrom.HasValue)
+                query = query.Where(s => s.Date >= filter.DateFrom.Value);
+
+            if (filter.DateTo.HasValue)
+                query = query.Where(s => s.Date <= filter.DateTo.Value);
+
+            if (filter.UserId.HasValue)
+                query = query.Where(s => s.UserId == filter.UserId.Value);
+
+            if (filter.GoalId.HasValue)
+                query = query.Where(s => s.GoalId == filter.GoalId.Value);
+
+            if (!string.IsNullOrEmpty(filter.Description))
+                query = query.Where(s => s.Description.Contains(filter.Description));
+
+            List<Saving> savings = [.. query.Include(s => s.User).Include(s => s.Goal).AsNoTracking()];
 
             IList<SavingVm> savingsVms = _mapper.Map<List<SavingVm>>(savings);
 
             return savingsVms;
         }
 
-        public IList<SavingVm> GetSavingsForYear(int year)
-        {
-            List<Saving> savings = _dbContext.Savings.Where(s => s.Date.Year == year).Include(s => s.User).Include(s => s.Goal).AsNoTracking().ToList();
-
-            IList<SavingVm> savingsVms = _mapper.Map<List<SavingVm>>(savings);
-
-            return savingsVms;
-        }
-
-        public IList<SavingVm> GetSavingsForMonth(int year, int month)
-        {
-            List<Saving> savings = _dbContext.Savings.Where(s => s.Date.Month == month && s.Date.Year == year).Include(s => s.User).Include(s => s.Goal).AsNoTracking().ToList();
-
-            IList<SavingVm> savingsVms = _mapper.Map<List<SavingVm>>(savings);
-
-            return savingsVms;
-        }
-
-        public async Task<Saving> AddSaving(AddOrEditSavingDto dto)
+        public async Task<SavingVm> AddSaving(AddOrEditSavingDto dto)
         {
             Goal? selectedGoal = _goalsService.GetGoal(dto.GoalId) ?? throw new Exception("Selected goal not exist.");
+            //ApplicationUser? selectedUser = _goalsService.GetGoal(dto.GoalId) ?? throw new Exception("Selected goal not exist.");
 
             Saving newSaving = new Saving
             {
@@ -74,12 +72,14 @@ namespace Savings_API.Services
             await _dbContext.Savings.AddAsync(newSaving);
             await _dbContext.SaveChangesAsync();
 
-            return newSaving;
+            SavingVm newSavingVm = _mapper.Map<SavingVm>(newSaving);
+
+            return newSavingVm;
         }
 
         public SavingVm? GetSavingVm(int savingId)
         {
-            Saving? saving = _dbContext.Savings.Find(savingId);
+            Saving? saving = _dbContext.Savings.Include(s => s.User).Include(s => s.Goal).AsNoTracking().FirstOrDefault(s => s.Id == savingId);
 
             SavingVm savingVm = _mapper.Map<SavingVm>(saving);
 
@@ -93,7 +93,7 @@ namespace Savings_API.Services
             return saving;
         }
 
-        public async Task<Saving> UpdateSaving(int savingId, AddOrEditSavingDto dto)
+        public async Task<SavingVm> UpdateSaving(int savingId, AddOrEditSavingDto dto)
         {
             Goal? selectedGoal = _goalsService.GetGoal(dto.GoalId) ?? throw new Exception("Selected goal not exist.");
 
@@ -107,10 +107,12 @@ namespace Savings_API.Services
             editedSaving.Description = dto.Description;
             editedSaving.Amount = dto.Amount;
             editedSaving.Date = dto.Date;
+            editedSaving.UserId = dto.UserId;
 
             await _dbContext.SaveChangesAsync();
+            SavingVm editedSavingVm = _mapper.Map<SavingVm>(editedSaving);
 
-            return editedSaving;
+            return editedSavingVm;
         }
 
         public async Task DeleteSaving(int savingId)