+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - App controller with health check and get api version.
 - Show API version on Swagger.
 - Set status of goal - Active, Unactive, Archive.
+- Filter Savings by description, userId, goalId, date.